Drop-in replacement for mutate

Provides supercharged version of mutate with group_by, order_by and aggregation over arbitrary window frame around a row.

Usage

mutate(x, ..., .by, .order_by, .frame, .index, .complete = FALSE)

Arguments

x

(data.frame)

...

expressions to be passed to mutate

.by

(expression, optional: Yes) columns to group by

.order_by

(expression, optional: Yes) columns to order by

.frame

(vector, optional: Yes) Vector of length 2 indicating the number of rows to consider before and after the current row. When argument .index is provided (typically a column of type date or datetime), before and after can be interval objects. See examples.

.index

(expression, optional: Yes) index column

.complete

(flag, default: FALSE) This will be passed to slider::slide / slider::slide_vec. Should the function be evaluated on complete windows only? If FALSE, the default, then partial computations will be allowed.

Value

data.frame

Details

A window function returns a value for every input row of a dataframe based on a group of rows (frame) in the neighborhood of the input row. This function implements computation over groups (partition_by in SQL) in a predefined order (order_by in SQL) across a neighborhood of rows (frame) defined by a (up, down) where

up/down are number of rows before and after the corresponding row

up/down are interval objects (ex: c(days(2), days(1)))

This implementation is inspired by spark's window API.

Implementation Details:

Iteration per row over the window is implemented using the versatile slider.

Application of a window aggregation can be optionally run in parallel over multiple groups (see argument .by) by setting a future parallel backend. This is implemented using furrr package.

function subsumes regular usecases of mutate

See also

Examples

library("magrittr")
# example 1
# Using iris dataset,
# compute cumulative mean of column `Sepal.Length`
# ordered by `Petal.Width` and `Sepal.Width` columns
# grouped by `Petal.Length` column

iris %>%
  mutate(sl_mean = mean(Sepal.Length),
         .order_by = c(Petal.Width, Sepal.Width),
         .by = Petal.Length,
         .frame = c(Inf, 0),
         ) %>%
  dplyr::slice_min(n = 3, Petal.Width, by = Species)
#> 
#> Attaching package: ‘purrr’
#> The following object is masked from ‘package:magrittr’:
#> 
#>     set_names
#> # A tibble: 15 × 6
#>    Petal.Length Sepal.Length Sepal.Width Petal.Width Species    sl_mean
#>           <dbl>        <dbl>       <dbl>       <dbl> <fct>        <dbl>
#>  1          1.1          4.3         3           0.1 setosa        4.3 
#>  2          1.4          4.9         3.6         0.1 setosa        4.85
#>  3          1.4          4.8         3           0.1 setosa        4.8 
#>  4          1.5          4.9         3.1         0.1 setosa        4.9 
#>  5          1.5          5.2         4.1         0.1 setosa        5.05
#>  6          4.1          5.8         2.7         1   versicolor    5.8 
#>  7          3.3          5           2.3         1   versicolor    5   
#>  8          3.3          4.9         2.4         1   versicolor    4.95
#>  9          3.7          5.5         2.4         1   versicolor    5.5 
#> 10          3.5          5           2           1   versicolor    5   
#> 11          4            6           2.2         1   versicolor    6   
#> 12          3.5          5.7         2.6         1   versicolor    5.35
#> 13          5.6          6.1         2.6         1.4 virginica     6.1 
#> 14          5            6           2.2         1.5 virginica     6   
#> 15          5.1          6.3         2.8         1.5 virginica     6.3 

# example 2
# Using a sample airquality dataset,
# compute mean temp over last seven days in the same month for every row

set.seed(101)
airquality %>%
  # create date column
  dplyr::mutate(date_col = lubridate::make_date(1973, Month, Day)) %>%
  # create gaps by removing some days
  dplyr::slice_sample(prop = 0.8) %>%
  dplyr::arrange(date_col) %>%
  # compute mean temperature over last seven days in the same month
  tidier::mutate(avg_temp_over_last_week = mean(Temp, na.rm = TRUE),
                 .order_by = Day,
                 .by = Month,
                 .frame = c(lubridate::days(7), # 7 days before current row
                            lubridate::days(-1) # do not include current row
                            ),
                 .index = date_col
                 )
#> # A tibble: 122 × 8
#>    Month Ozone Solar.R  Wind  Temp   Day date_col   avg_temp_over_last_week
#>    <int> <int>   <int> <dbl> <int> <int> <date>                       <dbl>
#>  1     7    64     175   4.6    83     5 1973-07-05                    83.5
#>  2     8    16      77   7.4    82     3 1973-08-03                    81  
#>  3     5    36     118   8      72     2 1973-05-02                   NaN  
#>  4     6    NA     286   8.6    78     1 1973-06-01                   NaN  
#>  5     8    78      NA   6.9    86     4 1973-08-04                    81.3
#>  6     5    12     149  12.6    74     3 1973-05-03                    72  
#>  7     9    78     197   5.1    92     2 1973-09-02                   NaN  
#>  8     6    NA     242  16.1    67     3 1973-06-03                    78  
#>  9     7    77     276   5.1    88     7 1973-07-07                    83.4
#> 10     8    35      NA   7.4    85     5 1973-08-05                    82.5
#> # ℹ 112 more rows